Separate buildings section from infrastructure
