+++ src/config/networks/hierarchy.ts
@@ -133,8 +133,4 @@ export const NETWORK_LAYERS_HIERARCHY: TreeNode<NetworkLayerData>[] = [
       },
     ],
   },
-  {
-    id: 'buildings',
-    label: 'Buildings',
-  },
 ];
